move did over to lib, and rename tests
