+++ tests/resources/ddo/ddo_sample_generated_2.json
@@ -1,24 +1,24 @@
 {
   "@context": "https://w3id.org/did/v1",
-  "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
-  "created": "2018-11-01 10:42:42.064975",
+  "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
+  "created": "2018-11-01 10:56:48.892156",
   "publicKey": [
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a#keys=1",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4#keys=1",
       "type": "RsaSignatureAuthentication2018",
-      "publicKeyPem": "-----BEGIN PUBLIC KEY-----\nMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDNkqAZ/QPC1MXeGqvbcpN/SElM\nt0Vp9VaBpkVh+vTa79viFWvUCz09GIvhxO5lVi1DyY4ktoXZVpHvQ4cY4tlP/YJ+\nFh7/vLAvxhHxDMRszc7cVtW4gy5FL+UVpTHEuUrGapcMsMKuXroTvDeN+yGg39I3\n3zG4oFYWYBIcpBcLWQIDAQAB\n-----END PUBLIC KEY-----",
-      "owner": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a#keys=1"
+      "publicKeyPem": "-----BEGIN PUBLIC KEY-----\nMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQC29GScCM/H8mYzXqjAW8Fckj5F\nN7IHJ4NaxHcfS8X9ksJFmo3ewsgopTqtsDAcVkMbnQJcRuIh2z2HyVkkSoAcltul\npuELRa2/2iax4Cst9ZuasNl0r/iMGe2Tl6oKlM4sZqV/Vg/YEpcRD5JkGsUC6zKt\nOnqThxRa2Lwub8xjYwIDAQAB\n-----END PUBLIC KEY-----",
+      "owner": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4#keys=1"
     }
   ],
   "authentication": [
     {
       "type": "RsaVerificationKey2018",
-      "publicKey": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a#keys=1"
+      "publicKey": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4#keys=1"
     }
   ],
   "service": [
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "Metadata",
       "serviceEndpoint": "http://myaquarius.org/api/v1/provider/assets/metadata/{did}",
       "metadata": {
@@ -75,32 +75,32 @@
       }
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "OpenIdConnectVersion1.0Service",
       "serviceEndpoint": "https://openid.example.com/"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "CredentialRepositoryService",
       "serviceEndpoint": "https://repository.example.com/service/8377464"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "XdiService",
       "serviceEndpoint": "https://xdi.example.com/8377464"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "HubService",
       "serviceEndpoint": "https://hub.example.com/.identity/did:op:0123456789abcdef/"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "MessagingService",
       "serviceEndpoint": "https://example.com/messages/8377464"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "SocialWebInboxService",
       "serviceEndpoint": "https://social.example.com/83hfh37dj",
       "description": "My public social inbox",
@@ -110,25 +110,25 @@
       }
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "BopsService",
       "serviceEndpoint": "https://bops.example.com/enterprise/"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "Consume",
       "serviceEndpoint": "http://mybrizo.org/api/v1/brizo/services/consume?pubKey=${pubKey}&serviceId={serviceId}&url={url}"
     },
     {
-      "id": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a",
+      "id": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4",
       "type": "Compute",
       "serviceEndpoint": "http://mybrizo.org/api/v1/brizo/services/compute?pubKey=${pubKey}&serviceId={serviceId}&algo={algo}&container={container}"
     }
   ],
   "proof": {
     "type": "RsaSignatureAuthentication2018",
-    "created": "2018-11-01 10:42:42.199713",
-    "creator": "did:op:c723a2774c2627a9fa49cbd703537284509b279732127d9685724c77972ed38a#keys=1",
-    "signatureValue": "b'L2qOJnVftPBdxBJWL7FXStrYhAUPo0KoBBIhQ6C7RKPqEDFhs33qnSPkHBeLniGwgZt+pcY/LIhcNpWJuf0RcVk13OcFOoRsu+M+HFZ3UnqZCTunpqL9buIUIfmOaFEDoBxgK14qINh/dmAaPLHYi/eU/zySjrINWqC9/S+aOPU='"
+    "created": "2018-11-01 10:56:48.988234",
+    "creator": "did:op:393819d44eef54da6e92ee384ace1c69878d3bdc34c00183755b046d258d62f4#keys=1",
+    "signatureValue": "b'BMPzr5dsvX4IDmvZ7C/hD+TuOpScjU3DqVSCvh+N6n7kCXykNKBqWckec9ItXfBRxXeNZapEEpqQ7ejUGX6qkiR+MorLz8/AW8gxNexHwWVS33adFN+TYf2YPmUeT8Yfqt+UdC5B2IEWATnqbPG5dnOo9rJmHEOxngvBrEX49Uw='"
   }
 }
