+++ tests/resources/ddo/ddo_sample_generated_2_private_key.pem
@@ -1,15 +1,15 @@
 -----BEGIN RSA PRIVATE KEY-----
-MIICXgIBAAKBgQDNkqAZ/QPC1MXeGqvbcpN/SElMt0Vp9VaBpkVh+vTa79viFWvU
-Cz09GIvhxO5lVi1DyY4ktoXZVpHvQ4cY4tlP/YJ+Fh7/vLAvxhHxDMRszc7cVtW4
-gy5FL+UVpTHEuUrGapcMsMKuXroTvDeN+yGg39I33zG4oFYWYBIcpBcLWQIDAQAB
-AoGABjNXvmjuCdyimjwqNlyKm+J5iYWWo6dJhHAYNCgeZhMpifR08+ih4HuiznzC
-5w4WH+8OqUdTGrbcgH/ELak7/jqPnLiBNvJnPA93CP67zTmaW33EHF6s1rfDFz4K
-IhJ3MvyywCUTCXnTPdfO4xNJNAepQuxW3WmkrVBEI4nm0k0CQQDakf9i+zyI27p4
-ydSEHJ1y9UkAlLjcb/lgOAfEsAMksjUbX0s2998J6D5HR+XhYTzc3X+6sdmj1ktN
-6pY8vfuFAkEA8MbSihhGcOaPsfW4ALN4iKbcp0db86SBiv+EKdK3C5vYABdunWpB
-zLauPhfZzu0PwN7Sa0078BZhoOApDq/mxQJBAJO6D9nPoveBQ4R2LavpR0W2oVPC
-MTvUdsFgvWdj64v24swQWV7ao6YnkUhOQQJW3XMORlFh+Y6xvQ35mveNFJ0CQQCO
-z98hObTOUnL4wx4YjFr58r7+GT1ZDkIoB9aEWA9qO02rADsD83qDyfcfrVXrfoDR
-jViRfIYIqb3OhibUJ1FhAkEAuzUzV94C47v2QRiWaJTRz97kqF8s5X3/GtRR1GYr
-T2qR57371EqS7Y36Gjw6rYIzUDDerCBlbAViqsDt5rk7xg==
+MIICXAIBAAKBgQC29GScCM/H8mYzXqjAW8Fckj5FN7IHJ4NaxHcfS8X9ksJFmo3e
+wsgopTqtsDAcVkMbnQJcRuIh2z2HyVkkSoAcltulpuELRa2/2iax4Cst9ZuasNl0
+r/iMGe2Tl6oKlM4sZqV/Vg/YEpcRD5JkGsUC6zKtOnqThxRa2Lwub8xjYwIDAQAB
+AoGAEoP0RKhNou4gsrUzPdhrXrt8KtQio7+3dvGoHpY0EOAP4ixsjC1JRXfHUvmg
+PNXsD7jZms4fp46BESNq7pXNlirQYKEmZmnJfK9k10S4JvLYazFaij4IvAGNGm2u
+OE7icNpPCJUmfK7yI+8w0BcCQL9ic5Mfyv6Ptt9mxLlEnukCQQC6UgqLuIy5WYev
+r1R1LsO7MNFtXRw3OBxEPKyUiUHlKPhrVU0oAQd+WzCwyAGZiFBJxf0KKPYPx/IJ
+Z58LRE5LAkEA+2Adi9zXFZYvI3qjs8Pmuby1ZP+6lZ+vYACHzLzui+nXES7ty/pn
+JtXPJJNVXoFQljk4OUxDxTM6FS/ZSWwwSQJBALas/1f0Dk0fwY/3qCfoPNGIUJ0Y
+QDTHl7czgcV0OZ4zqDScArTpJW62e6unvSHJ62YZbzmJ5Y6da+napa3HP7cCQGUZ
+dQipWz/tVSwZyPkl0wfko7fpOiuvO6OK0xM+8VoIRF0+2JjvLslpuJPnRbBqpfSl
+Mr2otKD9R2D6C5QCFTECQEmac2crMuy3coqymf190yFm63w26oh8DZrSVQXGM6qv
+0odwyWp4AT7vU0FJmU40C48OD8rcaV4fzw4yy+Rv7C0=
 -----END RSA PRIVATE KEY-----
