+++ tests/test_ddo.py
@@ -8,7 +8,7 @@ import pytest
 import secrets
 from web3 import Web3
 
-from did_ddo_lib.main import (
+from squid_py.did import (
     did_generate,
     did_generate_from_ddo,
     did_validate,
