+++ tests/test_did.py
@@ -7,7 +7,7 @@ import pytest
 import secrets
 from web3 import Web3
 
-from did_ddo_lib import (
+from squid_py.did import (
     did_generate,
     did_generate_from_ddo,
     did_parse,
